Fixed Scoping Issue
Apparently if you declare for(i=0... the scope for i is to large, and thus it carries through the stack. Instead by declaring the loop as for(var i =0... the scope is limited to that loop and does NOT carry through the stack. Which is what I want!

At this point the ai plays one piece at a time, but the moves are terrible. There is also an issue with the reset button.